requirements.txt
fastapi
pydantic
typing
openai
requests
pillow